continuing refactoring-change entity fields name to camel case

+++ src/main/java/com/example/book_club_proiect/services/UserService.java
@@ -3,9 +3,8 @@ package com.example.book_club_proiect.services;
 
 import com.example.book_club_proiect.dto.UserDTO;
 import com.example.book_club_proiect.mapper.UserMapper;
-import com.example.book_club_proiect.models.User;
-import com.example.book_club_proiect.repositories.RolesRepository;
-import com.example.book_club_proiect.repositories.UserRepository;
+import com.example.book_club_proiect.models.*;
+import com.example.book_club_proiect.repositories.*;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -18,12 +17,22 @@ public class UserService {
     private final UserMapper userMapper;
 
     private final UserRepository userRepository;
+    private final MyListingRepository myListingRepository;
+    private final BookOwnerRepository bookOwnerRepository;
+    private final WaitingListRepository waitingListRepository;
+    private final RentingTableRepository rentingTableRepository;
 
     private final RolesRepository rolesRepository;
 
-    public UserService(UserMapper userMapper, UserRepository userRepository, RolesRepository rolesRepository) {
+    public UserService(UserMapper userMapper, UserRepository userRepository, MyListingRepository myListingRepository,
+                       BookOwnerRepository bookOwnerRepository, WaitingListRepository waitingListRepository,
+                       RentingTableRepository rentingTableRepository, RolesRepository rolesRepository) {
         this.userMapper = userMapper;
         this.userRepository = userRepository;
+        this.myListingRepository = myListingRepository;
+        this.bookOwnerRepository = bookOwnerRepository;
+        this.waitingListRepository = waitingListRepository;
+        this.rentingTableRepository = rentingTableRepository;
         this.rolesRepository = rolesRepository;
     }
 
@@ -41,10 +50,10 @@ public class UserService {
     public User createUser(String first_name, String last_name, Integer user_age, String username, String user_email,
                            String user_password, Long role_id) {
         List<User> users = userRepository.findUsersByUsername(username).stream().toList();
-        if(users.size()>0){
+        if (users.size() > 0) {
 
-                throw new UnsupportedOperationException(
-                        "An user with this username  : " + username + " already exists");
+            throw new UnsupportedOperationException(
+                    "An user with this username  : " + username + " already exists");
         }
         User user = new User();
         user.setFirst_name(first_name);
@@ -59,8 +68,18 @@ public class UserService {
 
     }
 
-    public void deleteUserById(Long id) {
-        userRepository.deleteById(id);
+    public User deleteUserById(Long id) throws UnsupportedOperationException {
+        List<MyListing> listings = myListingRepository.findAllByReading_person(id).stream().toList();
+        List<BookOwner> bookOwners = bookOwnerRepository.findBookOwnersByUser_id(id);
+        List<WaitingList> waitingLists = waitingListRepository.findWaitingListsByUser_id(id);
+        List<RentingTable> rentingTableList =rentingTableRepository.findRentingTablesByBorrowed_by(id);
+        User user = userRepository.findById(id).get();
+        if(listings.size()>0 || bookOwners.size()>0 || waitingLists.size()>0 || rentingTableList.size()>0){
+            throw new UnsupportedOperationException(
+                    "You can not delete   : " + user.getUsername() + " because it is used in another places.");
+        }
+         userRepository.deleteById(id);
+        return user;
     }
 
     public User updateUser(Long id, UserDTO userDto) {
